endpoint: hide restore operations inside Endpoint package
The endpoint can be responsible for 'restoring' itself; the daemon does not have
to be aware of how it is done. This ensures that Endpoint implementation is not
leaked.

Signed-off by: Ian Vernon <ian@cilium.io>
